rename bloom_index/utils to laoder
