+++ src/query/storages/fuse/src/io/read/bloom_index/mod.rs
@@ -15,6 +15,6 @@
 
 mod column_reader;
 mod filter_reader;
-mod util;
+mod loader;
 
 pub use filter_reader::BloomFilterReader;
